refactor: root component
